Move Python wrapping code for root to pymod/
