+++ cpp/modmesh/CMakeLists.txt
@@ -37,6 +37,16 @@ set(MODMESH_ROOT_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/toggle/toggle.cpp
     CACHE FILEPATH "" FORCE)
 
+set(MODMESH_ROOT_PYMODHEADERS
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/root_pymod.hpp
+    CACHE FILEPATH "" FORCE)
+
+set(MODMESH_ROOT_PYMODSOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/root_modmesh.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_profile.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_Toggle.cpp
+    CACHE FILEPATH "" FORCE)
+
 set(MODMESH_DEVICE_HEADERS
     CACHE FILEPATH "" FORCE)
