+++ cpp/modmesh/pymod/root_modmesh.cpp
@@ -26,9 +26,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/python/python.hpp> // Must be the first include.
+#include <modmesh/pymod/root_pymod.hpp> // Must be the first include.
 #include <modmesh/buffer/pymod/buffer_pymod.hpp>
-#include <modmesh/python/wrapper/modmesh/modmesh.hpp>
 
 namespace modmesh
 {
